Move EventAction index.js to event.js
