+++ containers/EventForm/index.js
@@ -1,7 +1,7 @@
 // @flow
 import React from 'react'
 import { connect } from 'react-redux'
-import { createEvent } from '../../actions/Event/event'
+import { createEvent } from '../../actions/event'
 import EventFormComponent from '../../components/EventForm'
 
 type EventFormProps = {
